set up service to provide links. figured out content to element connection.
the reel component needed a boolean property to help deter which prop got the new displaying content. links will need to be in an achor tag as routelinks are specifically for angular navigation. next need to clean up landing animation and set up same content/element connection for the blurb component.