worked out last major found kink in reel animation!
the solution was simpler than expected. adding another condition in reel.scroll that uses a let  object to check if the current state is qued but opposite direction. if so, that change state to avoid having two consecutice qued states. found a glitch. when navagating back to the landing page something happens with the service subscription. qued and display props aren't getting populated correctly. my thinkin is this is what a 'data leak' looks like and a need to set up the ng end of lifecycle hook. Have to figure out what that is first...