add django project today.
totally beside myself excited to be working with django again. lots of questions. like with spa, am I limited to only serving one application? or can I have multple spa and otherwise apps in one django project? also installed djongo. have not set it up. another question. is it possible to do both MySQL and Mongo?  I hope its as flexible as I dream it is. Other wise flask sounds pretty good to me. it is  easier to use and I do thing I could string it up into a behemoth. but it wouldnt nearly be as pretty or slick as django. let django be flexible