landing page displaying bug content is due to nav buttons
need to do some major refactoring. the nav buttons need to end up on the parent component, landing-page, and the way both the blurb and reel component subscribe to the service has to change. this means I have to refactor the reel and landing page styling. might be a good time to ad some SASS. need to check if SASS is compatable with matrial module that I added. also want to refactor the styling of the wrap element so button is within the browser's viewing area.