Got Dialog behaving as I wanted.
Need to include an import in my style.css that gets the material theme I set when I installed material in this project. took alittle styling after that and everything looks good. next I will make another send resume button component for the resume page. after that I need to set up the dialog to send data to the grand parent component. Then I need to create and test an email html version of my resume. I wonder if there is a way to make the htmls content dynamic?