Added the other resume button component.
Looks like the dialog component can be hooked up to the email service and can grab the data enterd in the form just like any other form in angular. so I can off load the responsiblity of activating the service to the dialog and therefore not need to pass data to the parent or grand parent of the dialog component. next I need to hammer out the functionality of the Email Service methods.