reworked some of the functionality of the send resume button.
Now if the client has filled out the contact form when they click on a send resume button it will send a put request to /email with out openning up the dialog. need to come up with a success message that pops up when some one hits that button after they have filled out the contact form. maybe success dialog component with a setInterval to get it to leave the screen in a timely way.