All of my attempts to pre-populate the show page have failed.
MY PLAN:
  -add a resolver
It will give the content for the show project components onInit. The the project service will recieve the dll it will manipulate from the resolver via the last component to get loaded (i believe that to be the blurb).
  -add a projectApiService:
it will be in the one actually making the calls to the api. only the resolver should ever access this service.
 -test if animations transitions can be imported from another file.
 -add a hover animation for the landing page reel to aid in its UI.
  -USE THIS AS A CHANCE TO BRANCH:
branch it. then get it working in the branch. the merge and force yourself to resolve some merge conflicts.