Auto scroll working as expected.
Now to add the focus UI for the landing page reel. Hopefully the change of screen size wont screw with neighboring elements too much. Then I need to add the call to action component for show page. And I also need to test out how to modularize the animation transitions.