Reel hover/focus functionality mostly working as expected.
Still some positioning tweeks to make. mainly when the screen enlarges it spreads down from the top. The problem is, each of those screen elements has a different top or bottom value for its displayed position. I have to figure out some clever way to get it to grow from the middle of the elements position. I should take a look at the docs. next to implement an auto scroll for the show page reel. After that, start working on the call to action component for the landing page. Then, how about I put in some router animations? a simple slide to the left or right should work well.