Call to action component not working with my current impementation.
It is close. The problem is trying to get its state to go off at the right time and with the right string. I think either adding another state or some kind of logic to the current service state property. Maybe adding two functions, one to handle the state change to display, the other for the que change will do what I need.