Call to action working in reel.
I need to work out some animation glitches. Something todo with the frist animation state transition to the right (RightToQue). I set up a disable press to aviod the animation glitches that happen when you change state before the prior animation has completed. Need to add this same functionality with landing page. Now to add a sibling call to action component for the blurb of show page. I also want to add a nav component so that clients can navigate to any page from any page they are currently on. and then I need to set up toop tips for the site. then to start building and testing my database. I am at least over the hill with this project.