blurb call to action up and running as expected.
It seems the glitch with the show page animation and the call to action animation is caused by the when the service uses the cta state method. I need a way to keep it from triggering at the point that it's causing the glitch. Now to implement the disable nav press feature in the landing page.