Styling a little more buttoned up.
is it ever over? I still need to add some UI functionality to the buttons on the call to action component. Also need to encapsolate the nav  component so that it is easier to click on. Next thing to do as ad some clerifying tool tips. Then it is on to router animations. After all that I might button somethings up further on the client side. Or I might start building some of my models and testing in django.