Tooltips added and styled.
I might remove them from the landing page reel component. Still need add the button functionality for the call to action component. Nav component needs some jiggering. It is difficult to press in the show-page. Now on to add some router animation. I am going to have some clean up to do, but it seems the majority of the client functionality is set up. The last big thing to do is work the project api service to get data from the django backend. This means I need to build the data base models and the corrispoding route methods.