Router animation work as expexted for one route!
What a doozy. in the end I need to be more specific with my route changes. Also must of the tutorials I found deal with animating the root app's children. Basically the animation functionality I needed was mush simpler than implementations I found online. Now to refactor all routes to be activated via a method in the component rather than through a router link. After that, it's time to build out my backend. Models and Methods for django.