All route transitions working as expected.
Eventually, I will refactory all buttons to be their own component so that the route animation service an the router are only called in those components and not practiaclly every where else. nav sub menu needs to be pass project id data and some how that needs to be passed into the navToProject(). Also, nav is stll difficult to get menu in the show page. Button functionality in the call to action component still not working. The next steps are to merge this branch with the master, and handle all conflicts that arrise. Then I need to begin building out backend functionality. To start I need to refresh myself on Django's ORM. then I need to build those models. I should be able to test them in the python command line (I'll need a refresher there too). Once models look good, I will need to build out the route functionality for getting project data.