the get all and get one routes are working as expected.
Now on to wiring up the projects api service in the frontend. The service should format the json data into a dll and then hand that to the resolver or the landing page's projects service. Side note I will make the adimn and the storing of images into it their own micro services. Like I want to implement my own DevOps workflow to handle those two features. I should try and set up a chat with jason to see what insights he can give me on how to do that.