Phew. The site is getting and display content in an expected manor.
Did some refactor, including deleteling a couple of files I wasn't using, setting up a couple of new service dedicated to handling the dll for the show and landing pages, and renaming things for clerity. Now to set up the landing page to recieve it's content from the api. Eventually, I would like to set up the angular frontend to send csrf token. Also I would like to set up a micro app that just handles the image files for the site content. this micro service would working with a one page admin site that is also a micro service. The admin panel will be written in another frontend framework.