Landing page is now getting its content from the api!
Definitetly over the hill now. there are still some glitches I need to take care of. mainly in the show page. currently the animation for the call to action is triggered on the first nav press on the page. I need to re-do my resume. The call to action blurb needs some tweek. the reel sibling of that needs to be re-designed. it is ugly as all get out in it's current version. Also there is the matter of storing and serving the images for the site and then I need to style them. Speaking of the images, I also need to produce the copy for the site. oh boy. I need to get some help for that.