Landing page animations working as expected.
there is still needs to be some style adjustments for the blurb component. one of the elements is off from the other giving the animation an unwanted bounce. also the cta is triggered at the wrong time when it is first engaged. next I need to rewrite and build my resume.