perparing to launch the site.
it is not ready for production but I need to get something up. seeing as how I am using git I should be able to make changes and rebase on cloud as I need to update. I don't know for sure if that is possible. so this is really an expierenment.