Sensor server is now redirect to a different port based on device
So looking for a user-agent node package I found express-useragent. It seems to handle device detection completely. It seems now that I dont need 51Degrees. Now it is on to making a moblie version of the site. This might be a good chance to learn react rather start up another angular project for it.